// features/bankroll-tracker/types/contestEntry.ts
export type ContestEntry = {
  sport: string
  game_type: string
  entry_key: string
  entry: string
  contest_key: string
  contest_date_est: string
  place: string
  points: string
  winnings_non_ticket: string
  winnings_ticket: string
  contest_entries: string
  entry_fee: string
  prize_pool: string
  places_paid: string
  uploaded_at: string
}